feat(redux-enhancer): add a redux enhancer and return state after raising event
Adds a redux store enhancer that users can utilise to provide metadata that es-reduxed needs for
some capabilities. Adds ability to return the state that resulted from raising an event as a return
of the raise event function.

BREAKING CHANGE: Changed several user facing APIs in the provider and initialiser.