fix(queue): use redux event id to prevent out of order processing
If the event notifications from the event store were received out of order, there was a chance they
could be processed out of order and in duplicate. This fix prevents both occurrences by always
retrieving events based on the latest id processed in redux, rather than the event id in the queue.